modified questionnaire 16 and deleted unnecessary things
As the client's demand, the question 16 is optional, so I make the continue button clickable when the text box is empty, that user can just jump it this question if they don't want to answer this question. Besides, I delete the converage file as this is unnecessary and might make the code run slow.