count word
I have modified the few pages with text boxes so that the counter in the bottom right corner counts the number of words entered by the user, rather than all characters being counted. I have also added a progress bar to the finished question and fixed the problem where the text of the options in question 15 was going off screen.