Fix map and add timer
## Description
According to the game grid rearranged by chongqi, there is a small error which there should be no barrier near the right top guess box.
Beside, Based on the web game provided by client Kimberly that she had written, I found that the game would jump straight to the next stage if player stayed on the page for long time, so I added that feature. I've set the time to 5 min. 

## Type of Change
Please insert an X on the relevant option.

- [X] Bug fix (non-breaking change which fixes an issue)
- [X] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## How Has This Been Tested?
I set the timer to 10 second and see if it's work on the ios emulator.

### Tested On

**Desktop (please complete the following information)**
- OS: [IOS]
- Flutter Version (flutter doctor -v): [ 2.0.1]
- Version: [19]

**Smartphone (please complete the following information)**
- Device: [iphone 14 pro max]
- OS: [iOS]
- Version: [i1.8]

### Checklist:
Please insert an X on all that apply.

- [X] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [X] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [X] Any dependent changes have been merged and published in downstream modules