Add hyperlink for last page
## Description
According to the web the client provide us, I have add this paragraph
and add a hyperlink so that users can jump when they click on the URL.
Since adding a hyperlink requires adding a package of launch URL, you need to use flutter pub get to install this package. After installation the 'url_launcher' package was added to the 'pubspec.yaml' file, that's why the ".yaml" file was updated with some code.

## Type of Change
Please insert an X on the relevant option.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

## How Has This Been Tested?
If applicable, please state the tests you ran to verify your changes and its outcome.
This can include a screenshot to complement * passing * CI jobs.

### Tested On

**Desktop (please complete the following information)**
- OS: [i.e. Windows]
- Flutter Version (flutter doctor -v): [i.e. 2.0.1]
- Version: [i.e. 19]

**Smartphone (please complete the following information)**
- Device: [i.e. iPhone 13]
- OS: [i.e. iOS]
- Version: [i.e. 1.8]

### Checklist:
Please insert an X on all that apply.

- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] New and existing unit tests pass locally with my changes
- [ ] Any dependent changes have been merged and published in downstream modules